update text components
