+++ src/ui-components/basicInput/Text.js
@@ -1,3 +1,2 @@
-import {InputText, PasswordInText, TextArea} from './text/Text'
-
-export{InputText, PasswordInText, TextArea}
+import TextBox from './text/TextBox'
+export{TextBox}
