+++ src/ui-components/basicInput/text/Text.js
@@ -3,48 +3,7 @@ import PropTypes from 'prop-types';
 import './Style.css';
 
 
-export class InputText extends React.Component {
-    state = {
-        keyword: this.props.value
-    }
 
-    static getDerivedStateFromProps(props, state) {
-        return {
-            keyword: props.value
-        }
-    }
-
-    _inputChangedHandler(event) {
-        this.setState({ keyword: event.target.value })
-        this.props.onChangeText(event.target.value)
-    }
-
-    render() {
-        const {
-            placeholder = "Hint text"
-        } = this.props
-        return (
-            <input
-                type="text"
-                className='InText'
-                placeholder={placeholder}
-                value={this.state.keyword}
-                onChange={(event) => this._inputChangedHandler(event)}
-            >
-            </input>
-        )
-    }
-}
-
-InputText.propTypes = {
-    disabled: PropTypes.bool,
-    onChangeText: PropTypes.func
-};
-
-InputText.defaultProps = {
-    disabled: false,
-    onChangeText: noAction
-};
 
 export class PasswordInText extends React.Component {
     state = {
